Assignment 3: Fundamental Layout Commit 3
Part 4: Added all images to an images folder and began editing the CSS selectors.
First, I added the display: flex, justify-content: space-between, and align-items: center configurations to the nav ul selector to display the navigation items in an evenly spaced row.
Next, I added sticky positioning to the nav selector so that the nav bar is always seen at the TOP when scrolling through the page.