Assignment 3: Fundamental Layout Commit 6
Part 4: Added a display grid property to the .photos selector as well as splitting the section into two even columns using the grid-template-columns property with a value of 1fr 1fr. I added a gap of 1 pixel to the selector as well.
I then did some troubleshooting (because the images had large gaps between them) and found my mistake of having the float and margin values in the wrong selector (img). After moving them to the .feature selector, the mistake was corrected.