Assignment 4: Story Commit 4
Part 1: Began step 3 by first initializing the newStory variable (to allow for a new story on each click).
Initialized item variables for each array to store a random item from each (randomValueFromArray).
Replaced the values in the story with the random items using .replaceAll (for insertx) and .replace (for inserty and z).
Finally, completed the function to replace Bob's name with a user-entered name (if applicable) using .replace().