Assignment 4: Story Commit 5
Part 1: Completed the conversion function by adding conversion calculations to the weight and temperature variables and concatenating into a template string.
Used the .replace() method to replace the instances in the story string with the corresponding variables.
Set story.textContent equal to the modified newStory variable to display the silly story complete  with random values.