Assignment 4: Balls Commit 4
Part 4: Began initializing the EvilCircle class. The class extends Shape, is only passed x and y args that are passed up to the Shape superclass along with velX and velY values of 20.
The colour and size were set, and the provided event listener was added to the constructor.
Finally, the draw method was implemented, using the same values as the one in Ball, with changes to stroke/strokeStyle and lineWidth.