Assignment 4: Balls Commit 5
Part 4: Created the checkBounds() and collisionDetect() methods in the EvilCircle class using the  corresponding methods in the Ball class and making necessary edits.
Ensured that if the evil circle overlaps with a ball, the ball is set to not exist.
Added comments for clarity throughout the updated classes and methods.