Assignment 4: Balls Commit 7
Part 4: Began by adding the "Ball Count: " paragraph in the HTML file.
Then added the paragraph styling to the CSS file.
Finally, added a reference to the ball count paragraph in the JavaScript file along with initializing a ballCount variable to 0. In the collisionDetect() method of EvilCircle, add logic to decrement ballCount when ball no longer exists (collides with evil circle) and update text context of paragraph.
After ball is pushed to the Balls array, add counter logic to increment ballCount (up to 25) and update the text content of the paragraph.